requirements.txt
fastapi
databases
sqlalchemy
asyncpg
python-dotenv
uvicorn
boto3
psycopg2-binary
